+++ lib/one_context.dart
@@ -1,3 +1,8 @@
 library one_context;
 
 export 'package:one_context/src/controllers/one_context.dart' show OneContext;
+export 'package:one_context/src/components/one_hot_reload.dart'
+    show OneHotReload, OneHotReloadBuilder;
+export 'package:one_context/src/controllers/one_theme_controller.dart'
+    show OneThemeController, OneThemeChangerEvent;
+export 'package:one_context/src/controllers/one_platform.dart' show OnePlatform;