+++ lib/src/components/one_context_widget.dart
@@ -1,5 +1,5 @@
-import 'package:one_context/src/controllers/one_context.dart';
 import 'package:flutter/material.dart';
+import 'package:one_context/src/controllers/one_context.dart';
 
 class OneContextWidget extends StatefulWidget {
   final Widget child;